2979054: Support for 0..n plural forms instead of only 1.
The number of plural forms is hard-coded in the filter for all languages.
(If a language is missing, the default is 1 plural form)
If for a given language the number is n, then the translated po-file contains
msgstr[0] ".." till msgstr[n] ".." lines. The number of msgstr[x] lines
in the source file does not matter. if the source contains more lines than n,
then the extra lines are ignored (and not aligned).